Split VaultManager into focused sub-actors
Priority 4 from vault architecture review.

Extract two internal actors from VaultManager:

VaultKeyCoordinator (~270 lines):
- GRDB observation for unshared inboxes
- Individual key sharing (DeviceKeyShare)
- Bulk key sharing (DeviceKeyBundle) via shareAllKeys()
- Key import (importKeyShare, importKeyBundle)
- Owns inboxesBeingShared tracking set

VaultDeviceManager (~150 lines):
- Device list/add/remove operations
- Sync vault group members to database
- Device name resolution from vault messages
- Pending peer device name tracking for pairing

VaultManager (~280 lines, was ~350 + ~200 extension):
- Thin coordinator that creates sub-actors and routes callbacks
- Bootstrap/lifecycle (connect, pause, resume, disconnect)
- VaultClientDelegate routing to appropriate sub-actor
- Conversation deletion broadcast
- Unpair/self-removal cleanup

Pairing stays as VaultManager extension for now — it's tightly
coupled to VaultClient DM streams and delegate callbacks. Extracting
it requires VaultClient refactoring that would be a separate PR.

Also add VaultClient.isConnected convenience property.
