+++ ConvosCore/Sources/ConvosCore/Vault/VaultManager.swift
@@ -42,8 +42,8 @@ public actor VaultManager {
     let databaseReader: any DatabaseReader
     var databaseWriter: (any DatabaseWriter)?
     let deviceName: String
-    var pendingPeerDeviceNames: [String: String] = [:]
-    var inboxObservationCancellable: AnyCancellable?
+    let keyCoordinator: VaultKeyCoordinator
+    let deviceManager: VaultDeviceManager
 
     public weak var delegate: (any VaultManagerDelegate)?
     public weak var eventHandler: (any VaultEventHandler)?
@@ -51,15 +51,11 @@ public actor VaultManager {
     var dmStreamTask: Task<Void, Never>?
     var activePairingSlug: String?
     var joinerDmStreamTask: Task<Void, Never>?
-    var inboxesBeingShared: Set<String> = []
 
     public private(set) var bootstrapState: VaultBootstrapState = .notStarted
 
     public var isConnected: Bool {
-        get async {
-            if case .connected = await vaultClient.currentState { return true }
-            return false
-        }
+        get async { await vaultClient.isConnected }
     }
 
     public var hasMultipleDevices: Bool {
@@ -77,11 +73,23 @@ public actor VaultManager {
         databaseReader: any DatabaseReader,
         deviceName: String
     ) {
-        self.vaultClient = VaultClient()
+        let client = VaultClient()
+        self.vaultClient = client
         self.identityStore = identityStore
         self.vaultKeyStore = vaultKeyStore
         self.databaseReader = databaseReader
         self.deviceName = deviceName
+        self.keyCoordinator = VaultKeyCoordinator(
+            vaultClient: client,
+            identityStore: identityStore,
+            databaseReader: databaseReader,
+            deviceName: deviceName
+        )
+        self.deviceManager = VaultDeviceManager(
+            vaultClient: client,
+            databaseReader: databaseReader,
+            deviceName: deviceName
+        )
     }
 
     public func setDelegate(_ delegate: any VaultManagerDelegate) {
@@ -90,12 +98,13 @@ public actor VaultManager {
 
     public func setEventHandler(_ handler: any VaultEventHandler) {
         self.eventHandler = handler
+        Task { await keyCoordinator.setEventHandler(handler) }
     }
 
     public func connect(signingKey: SigningKey, options: ClientOptions) async throws {
         await vaultClient.setDelegate(self)
         try await vaultClient.connect(signingKey: signingKey, options: options)
-        await syncDevicesToDatabase()
+        await deviceManager.syncToDatabase()
     }
 
     public func bootstrapVault(
@@ -103,6 +112,8 @@ public actor VaultManager {
         environment: AppEnvironment
     ) async {
         self.databaseWriter = databaseWriter
+        await keyCoordinator.setDatabaseWriter(databaseWriter)
+        await deviceManager.setDatabaseWriter(databaseWriter)
 
         guard let vaultKeyStore else {
             Log.warning("No VaultKeyStore configured, skipping vault bootstrap")
@@ -148,7 +159,7 @@ public actor VaultManager {
             try await inboxWriter.save(inboxId: inboxId, clientId: installationId, isVault: true)
             bootstrapState = .ready
             Log.info("Vault bootstrapped: inboxId=\(inboxId)")
-            startObservingInboxes()
+            await keyCoordinator.startObservingInboxes()
         } catch {
             let message = "Failed to bootstrap vault: \(error)"
             bootstrapState = .failed(message)
@@ -180,94 +191,30 @@ public actor VaultManager {
 
     public func resume() async {
         await vaultClient.resume()
-        await syncDevicesToDatabase()
+        await deviceManager.syncToDatabase()
     }
 
-    // MARK: - Key Sharing
+    // MARK: - Key Sharing (delegates to VaultKeyCoordinator)
 
     public func shareAllKeys() async throws {
-        guard let installationId = await vaultClient.installationId else {
-            throw VaultClientError.notConnected
-        }
-
-        let inboxRows = try await databaseReader.read { db -> [InboxConversationRow] in
-            let sql = """
-                SELECT i.inboxId, i.clientId, c.id as conversationId
-                FROM inbox i
-                INNER JOIN conversation c ON c.clientId = i.clientId
-                    AND c.id NOT LIKE 'draft-%'
-                    AND c.isUnused = 0
-                WHERE i.isVault = 0
-                """
-            return try Row.fetchAll(db, sql: sql).map { row in
-                InboxConversationRow(
-                    inboxId: row["inboxId"],
-                    clientId: row["clientId"],
-                    conversationId: row["conversationId"] ?? ""
-                )
-            }
-        }
-
-        var keys: [DeviceKeyEntry] = []
-        for item in inboxRows {
-            guard let identity = try? await identityStore.identity(for: item.inboxId) else { continue }
-            keys.append(DeviceKeyEntry(
-                conversationId: item.conversationId,
-                inboxId: item.inboxId,
-                clientId: item.clientId,
-                privateKeyData: Data(identity.keys.privateKey.secp256K1.bytes),
-                databaseKey: identity.keys.databaseKey
-            ))
-        }
-
-        let peerNames = pendingPeerDeviceNames.isEmpty ? nil : pendingPeerDeviceNames
-        let bundle = DeviceKeyBundleContent(
-            keys: keys,
-            senderInstallationId: installationId,
-            senderDeviceName: deviceName,
-            peerDeviceNames: peerNames
-        )
-
-        try await vaultClient.send(bundle, codec: DeviceKeyBundleCodec())
-        pendingPeerDeviceNames.removeAll()
-        if let databaseWriter, !inboxRows.isEmpty {
-            let inboxIds = inboxRows.map { $0.inboxId }
-            try? await databaseWriter.write { db in
-                for id in inboxIds {
-                    try db.execute(
-                        sql: "UPDATE inbox SET sharedToVault = 1 WHERE inboxId = ?",
-                        arguments: [id]
-                    )
-                }
-            }
-        }
+        let peerNames = await deviceManager.pendingPeerDeviceNames
+        try await keyCoordinator.shareAllKeys(pendingPeerDeviceNames: peerNames)
+        await deviceManager.clearPendingPeerNames()
     }
 
-    // MARK: - Device Management
+    // MARK: - Device Management (delegates to VaultDeviceManager)
 
     public func listDevices() async throws -> [VaultDevice] {
-        let dbDevices = try VaultDeviceRepository(dbReader: databaseReader).fetchAll()
-        if dbDevices.isEmpty {
-            return [VaultDevice(inboxId: await vaultInboxId ?? "self", name: deviceName, isCurrentDevice: true)]
-        }
-        return dbDevices.map { VaultDevice(inboxId: $0.inboxId, name: $0.name, isCurrentDevice: $0.isCurrentDevice) }
+        try await deviceManager.listDevices()
     }
 
     public func addMember(inboxId: String) async throws {
-        try await vaultClient.addMember(inboxId: inboxId)
-        await syncDevicesToDatabase()
-        await checkUnsharedInboxes()
+        try await deviceManager.addMember(inboxId: inboxId)
+        await keyCoordinator.checkUnsharedInboxes()
     }
 
     public func removeDevice(inboxId: String) async throws {
-        let removal = DeviceRemovedContent(
-            removedInboxId: inboxId,
-            reason: .userRemoved
-        )
-
-        try await vaultClient.send(removal, codec: DeviceRemovedCodec())
-        try await vaultClient.removeMember(inboxId: inboxId)
-        await syncDevicesToDatabase()
+        try await deviceManager.removeDevice(inboxId: inboxId)
     }
 
     public func broadcastConversationDeleted(inboxId: String, clientId: String) async {
@@ -312,8 +259,7 @@ public actor VaultManager {
             try? await writer.replaceAll([])
         }
 
-        inboxObservationCancellable?.cancel()
-        inboxObservationCancellable = nil
+        await keyCoordinator.stopObserving()
         bootstrapState = .notStarted
     }
 
@@ -323,88 +269,12 @@ public actor VaultManager {
         await cleanupLocalVaultState(inboxId: selfInboxId)
     }
 
-    func syncDevicesToDatabase() async {
-        guard let databaseWriter else { return }
-        guard let selfInboxId = await vaultInboxId else { return }
-
-        do {
-            let members = try await vaultClient.members()
-
-            let existingNames = try VaultDeviceRepository(dbReader: databaseReader)
-                .fetchAll()
-                .reduce(into: [String: String]()) { $0[$1.inboxId] = $1.name }
-
-            let unknownMembers = members.filter { member in
-                member.inboxId != selfInboxId
-                    && pendingPeerDeviceNames[member.inboxId] == nil
-                    && existingNames[member.inboxId] == nil
-            }
-
-            var messageNames: [String: String] = [:]
-            if !unknownMembers.isEmpty {
-                messageNames = try await loadDeviceNames()
-            }
-
-            let writer = VaultDeviceWriter(dbWriter: databaseWriter)
-
-            let devices = members.map { member in
-                let isSelf = member.inboxId == selfInboxId
-                let name: String
-                if isSelf {
-                    name = deviceName
-                } else {
-                    name = pendingPeerDeviceNames[member.inboxId]
-                        ?? existingNames[member.inboxId]
-                        ?? messageNames[member.inboxId]
-                        ?? "Unknown device"
-                }
-                return DBVaultDevice(
-                    inboxId: member.inboxId,
-                    name: name,
-                    isCurrentDevice: isSelf
-                )
-            }
-
-            try await writer.replaceAll(devices)
-        } catch {
-            Log.error("Failed to sync vault devices to database: \(error)")
-        }
-    }
-
     public static var preview: VaultManager {
         VaultManager(
             identityStore: MockKeychainIdentityStore(),
             databaseReader: try! DatabaseQueue(), // swiftlint:disable:this force_try
             deviceName: "Preview Device"
         )
-    }
-
-    struct InboxConversationRow {
-        let inboxId: String
-        let clientId: String
-        let conversationId: String
-    }
-
-    private func loadDeviceNames() async -> [String: String] {
-        guard let messages = try? await vaultClient.vaultGroupMessages() else { return [:] }
-
-        var names: [String: String] = [:]
-        for message in messages {
-            if let bundle: DeviceKeyBundleContent = try? message.content() {
-                if let name = bundle.senderDeviceName {
-                    names[message.senderInboxId] = name
-                }
-                if let peers = bundle.peerDeviceNames {
-                    for (inboxId, peerName) in peers {
-                        names[inboxId] = peerName
-                    }
-                }
-            } else if let share: DeviceKeyShareContent = try? message.content(),
-                      let name = share.senderDeviceName {
-                names[message.senderInboxId] = name
-            }
-        }
-        return names
     }
 }
 
@@ -413,15 +283,15 @@ public actor VaultManager {
 extension VaultManager: VaultClientDelegate {
     nonisolated public func vaultClient(_ client: VaultClient, didReceiveKeyBundle bundle: DeviceKeyBundleContent, from senderInboxId: String) {
         Task {
-            await self.syncDevicesToDatabase()
-            await self.importKeyBundle(bundle)
+            await self.deviceManager.syncToDatabase()
+            await self.keyCoordinator.importKeyBundle(bundle)
         }
     }
 
     nonisolated public func vaultClient(_ client: VaultClient, didReceiveKeyShare share: DeviceKeyShareContent, from senderInboxId: String) {
         Task {
-            await self.syncDevicesToDatabase()
-            await self.importKeyShare(share)
+            await self.deviceManager.syncToDatabase()
+            await self.keyCoordinator.importKeyShare(share)
         }
     }
 
@@ -432,7 +302,7 @@ extension VaultManager: VaultClientDelegate {
             if removal.removedInboxId == selfInboxId {
                 await self.handleSelfRemoved()
             } else {
-                await self.syncDevicesToDatabase()
+                await self.deviceManager.syncToDatabase()
             }
         }
     }
